move files into directories
