+++ src/renderer/navigation/NavBar.tsx
