+++ src/renderer/navigation/NavBarItem.tsx
